BACKEND_APP/pyproject.toml
[tool.poetry]
name = "emir-stroy"
version = "0.1.0"
description = ""
authors = ["Baiel <jyrgalbekuulubaiel@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.2"
djangorestframework = "^3.16.0"
django-cors-headers = "^4.7.0"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.1.0"
django-filter = "^25.1"
djangorestframework-simplejwt = "^5.5.0"
pandas = "^2.2.3"
numpy = "^2.2.5"
drf-yasg = "^1.21.10"
django-extensions = "^4.1"
django-unfold = "^0.56.0"
celery = "^5.5.2"
aiogram = "^3.4"


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
flake8 = "^7.2.0"
pytest = "^8.3.5"
pytest-django = "^4.11.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
